Move TestBase into bases package
